Small tidy-up in PersistentJournal implementation.
git-svn-id: svn+ssh://192.168.1.25/home/matthew/Repos/repo_01/phatbooks/trunk@2443 ac3cbfda-f6ac-4daa-b570-d46c5f00fc98

+++ src/persistent_journal.cpp
@@ -151,7 +151,7 @@ PersistentJournal::save_new_journal_core()
 	);
 	statement.bind
 	(	":transaction_type_id",
-		static_cast<int>(static_cast<int>(Journal::do_get_transaction_type()))
+		static_cast<int>(Journal::do_get_transaction_type())
 	);
 	statement.bind(":comment", wx_to_std8(Journal::do_get_comment()));
 	statement.step_final();
